feat: implement WorkItem into focus recent section
